requirements.txt
alembic==0.8.8
aniso8601==1.2.0
appy.pod==0.2.1
appy.shared==0.1.0
autoenv==1.0.0
bleach==1.4.3
blinker==1.4
chardet==2.3.0
click==6.6
colorama==0.3.7
decorator==4.0.6
docutils==0.12
dominate==2.2.1
enum34==1.1.6
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-HTTPAuth==3.2.1
Flask-JWT==0.3.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Mako==0.4
Flask-Migrate==2.0.0
Flask-Moment==0.5.1
Flask-OAuthlib==0.9.3
Flask-PageDown==0.2.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
ForgeryPy==0.1
futures==3.0.5
grpcio==1.0.1
html5lib==0.9999999
httpie==0.9.6
ibm-db==2.0.7
ibm-db-sa==0.3.2
ipython==2.4.1
itsdangerous==0.24
Jinja2==2.8
JPype1==0.6.1
lamson==1.3.4
lockfile==0.12.2
lxml==3.6.4
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
nose==1.3.7
numpy==1.11.2
oauthlib==2.0.0
passlib==1.6.5
pbr==1.10.0
pexpect==4.0.1
protobuf==3.1.0.post1
ptyprocess==0.5
py==1.4.31
pydevd==0.0.6
Pygments==2.1.3
PyJWT==1.4.2
pyreadline==2.1
pytest==3.0.3
python-daemon==2.1.1
python-dateutil==2.6.0
python-editor==1.0.1
python-modargs==1.7
pytz==2016.7
pyxdg==0.25
PyYAML==3.12
requests==2.11.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.1
stevedore==1.17.1
suds-jurko==0.6
suds-lxml==0.7.dev1
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
visitor==0.1.3
Werkzeug==0.11.11
WTForms==2.1
xmltodict==0.10.2
